Update Flask to v2.0 and refactor around typing issues
