+++ pyproject.toml
@@ -16,7 +16,7 @@ Pillow = "^8.0.1"
 arrow = "^0.17.0"
 atomicwrites = "^1.4.0"
 awesome-slugify = "^1.6.5"
-Flask = "^1.1.2"
+Flask = "^2.0.0"
 Flask-Caching = "^1.9.0"
 pony = "^0.7.14"
 Pygments = "^2.7.3"
